refactor(ui): Move CopyToClipboard to common
