+++ apps/clearance_ui/src/components/common/CopyToClipboard.tsx
